% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_authority.R
\name{mod_authority_server}
\alias{mod_authority_server}
\title{Authority Profile Module Server}
\usage{
mod_authority_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Authority Profile Module Server
}
