% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_evidence.R
\name{mod_evidence_server}
\alias{mod_evidence_server}
\title{Evidence Explorer Module Server}
\usage{
mod_evidence_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Evidence Explorer Module Server
}
