% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicators.R
\name{mod_indicators_server}
\alias{mod_indicators_server}
\title{Indicators Module Server}
\usage{
mod_indicators_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Indicators Module Server
}
