% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicators.R
\name{mod_indicators_ui}
\alias{mod_indicators_ui}
\title{Indicators Module UI}
\usage{
mod_indicators_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Displays indicator-level scores and a breakdown of contributions.
}
