% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pillar_detail.R
\name{mod_pillar_detail_server}
\alias{mod_pillar_detail_server}
\title{Pillar Detail Module Server}
\usage{
mod_pillar_detail_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Pillar Detail Module Server
}
