% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_risk.R
\name{mod_risk_server}
\alias{mod_risk_server}
\title{Risk Register Module Server}
\usage{
mod_risk_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Risk Register Module Server
}
