% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sensitivity.R
\name{mod_sensitivity_server}
\alias{mod_sensitivity_server}
\title{Sensitivity Analysis Module Server}
\usage{
mod_sensitivity_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Sensitivity Analysis Module Server
}
