% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sensitivity.R
\name{mod_sensitivity_ui}
\alias{mod_sensitivity_ui}
\title{Sensitivity Analysis Module UI}
\usage{
mod_sensitivity_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Allows users to explore SCI sensitivity under different weight scenarios.
}
