% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_targets.R
\name{mod_targets_server}
\alias{mod_targets_server}
\title{Readiness Targets Module Server}
\usage{
mod_targets_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Readiness Targets Module Server
}
