% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trace.R
\name{mod_trace_server}
\alias{mod_trace_server}
\title{Traceability Module Server}
\usage{
mod_trace_server(id, evidence_rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{evidence_rv}{A reactive returning a validated evidence data.frame.}
}
\value{
Invisible \code{NULL}.
}
\description{
Traceability Module Server
}
