% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trace.R
\name{mod_trace_ui}
\alias{mod_trace_ui}
\title{Traceability Module UI}
\usage{
mod_trace_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Displays traceability indicator scores and evidence coverage.
}
